Make renderWrapper.tsx compile
The typescript compiler was trying to tell me that I had an unused
import, and that the import couldn't be resolved. But happo was hiding
that error so I had to dig in to the happo.io code and add some
console logging.

The bug is taken care of in https://github.com/happo/happo.io/pull/35.
This commit will just bring the project to a good state again.